requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
asarPy==1.0.1
Cython==3.1.3
Mako==1.3.10.dev0
Markdown==3.8.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==25.0
Pygments==2.19.2
setuptools==80.9.0
six==1.17.0
uv==0.8.13
wheel==0.46.1
